udpate _bench

+++ _bench/fib/fib.jl
@@ -1,6 +1,6 @@
 #!/usr/bin/env julia
 
-fib(x) = x > 1 ? fib(x - 1) + fib(x - 2) : 1
+fib(x::UInt64) = x > 1 ? fib(x - 1) + fib(x - 2) : 1
 
-n = 40
+n::UInt64 = 40
 println(fib(n))
