+++ _bench/fib/fib.rs
@@ -1,4 +1,4 @@
-fn fib(n: usize) -> usize {
+fn fib(n: u64) -> u64 {
     if n < 2 {
         return 1;
     } else {
